Allow downloading stage1 from CI
- Check if `compiler/` is unmodified
- Check for untracked files in `compiler/`

  Adding build.rs can change the build inputs even if no existing files
  are changed :(

- Use the existing rust-lang/rust remote instead of adding a new one

  This allows the logic to work in read-only files systems, like GitHub
  actions.

- Require opting-in to downloading stage1
- Add caching for stage1 sysroot
- Download rustc-dev
- Don't rebuild stage1 std unnecessarily

+++ src/bootstrap/builder.rs
@@ -629,8 +629,10 @@ impl<'a> Builder<'a> {
                     .join("rustlib")
                     .join(self.target.triple)
                     .join("lib");
-                let _ = fs::remove_dir_all(&sysroot);
-                t!(fs::create_dir_all(&sysroot));
+                if !crate::use_cached_rustc(&self.compiler) {
+                    let _ = fs::remove_dir_all(&sysroot);
+                    t!(fs::create_dir_all(&sysroot));
+                }
                 INTERNER.intern_path(sysroot)
             }
         }
